% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{Ic}
\alias{Ic}
\title{Group Cohesion Index For Radial Arm Maze}
\usage{
Ic(df)
}
\arguments{
\item{df}{A data frame containing the number of animals in each zone of the
radial arm maze. The data frame should be organized as follows: each column
must correspond to a zone of the radial arm maze and each row must
correspond to a different observation.}
}
\value{
A vector of length \code{nrow(df)}.
}
\description{
This function calculates the cohesion index of groups of animals
 navigating a radial arm maze. Maximum cohesion occurs when all group members
 are in the same zone of the radial arm maze; minimum cohesion occurs when
 group members are equally spread between all the zones of the radial arm
 maze.
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
